Remove relative paths
Relative paths will work when using attachments but not
when using file references. Even if the relative path
works in JavaScript the base path won't be the same when
the relative path is used by PSaaS Builder. This will
cause errors in Builder or empty files in the FGM.
Relative paths could also be converted to absolute before
using them using path.resolve.