Fix fuelmap name
The fuelmap was being added to the Prometheus document as
elevation.prj. This meant there were duplicate elevation.prj
files (the fuel type one being overriden by the actual projection
file during execution) and no fuel map. Prometheus validation
failed but the message saying so was never received because of
the issue resolved in commit
3dccb38a5d6f710b49754e3007cbee21240a005d.